backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-unfold==0.89.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
sqlparse==0.5.5
tzdata==2026.1
python-dotenv==1.0.0
